+++ src/main/java/subway/application/feecalculator/FeeCalculator.java
@@ -1,4 +1,4 @@
-package subway.application.feepolicy;
+package subway.application.feecalculator;
 
 public interface FeeCalculator {
